requirements.txt
# Manga Translator - Requirements
# Optimized for Intel CPU (i7 Gen 11) with Combined Architecture
# Supports: Japanese (MangaOCR), Korean/Chinese/English (EasyOCR)
# Advanced Features: SAM2 segmentation, Flux AI inpainting

# ==============================================================================
# Core dependencies
# ==============================================================================
# numpy 2.x required for Python 3.13+
numpy>=2.0.0
opencv-python>=4.9.0.80
pillow>=10.3.0
tqdm>=4.66.2
scipy>=1.10.0

# ==============================================================================
# ML Models - Detection (YOLO)
# ==============================================================================
ultralytics>=8.1.43
torch>=2.0.1
torchvision>=0.15.2

# ==============================================================================
# ML Models - Japanese OCR
# ==============================================================================
manga-ocr>=0.1.11
sentencepiece>=0.2.0
huggingface-hub>=0.22.2
safetensors>=0.4.2

# ==============================================================================
# ML Models - Multi-language OCR (Korean, Chinese, English)
# ==============================================================================
easyocr>=1.7.0

# ==============================================================================
# Advanced Detection - SAM2 Segmentation (quality/premium mode)
# ==============================================================================
# Option 1: Official SAM2 (recommended for CUDA GPU)
# sam2>=1.0.0

# Option 2: Transformers SAM (works on CPU/Intel GPU)
transformers>=4.35.0

# ==============================================================================
# Advanced Inpainting - Flux AI (premium mode)
# ==============================================================================
# For colored bubble inpainting with AI
diffusers>=0.25.0
accelerate>=0.25.0

# ==============================================================================
# Translation API
# ==============================================================================
httpx>=0.27.0
python-dotenv>=1.0.1

# ==============================================================================
# Web Framework
# ==============================================================================
Flask>=2.2.5
flask-cors>=4.0.0
Werkzeug>=3.0.1
gunicorn>=21.2.0
gradio>=4.0.0

# ==============================================================================
# Security
# ==============================================================================
cryptography>=42.0.4

# ==============================================================================
# Intel CPU/GPU Optimization (recommended for Intel i7)
# ==============================================================================
openvino>=2024.0.0

# ==============================================================================
# Legacy translator support (optional)
# ==============================================================================
deep-translator>=1.11.4
translators>=5.9.1
